Rename all added functions to avoid name clashes (#4)
In order to avoid any problem with duplicate names, all the added
functions now use both their package and function names. (E.g.,
filepath.Rel becomes filepathrel.) A partial exception is
wordwrap.WrapString. This function is provided both as wordwrapwrapstring
(following the default pattern) and as wrapstring (for convenience).

I've also updated the documentation in txtartmpl.go and README.md and
the example in hello.txtar.